requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2020.6.20
click==7.1.2
cloudinary==1.22.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
Django==3.1
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-media-pil==1.0.0
django-registration==3.1
django-tinymce==3.3.0
djangorestframework==3.12.4
gunicorn==20.0.4
install==1.3.4
Pillow==8.3.1
psycopg2==2.8.6
pycrop==1.0.1
python-decouple==3.3
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
whitenoise==5.2.0
yapf==0.30.0